refactor: move methods from mixin to hoc
